+++ NEWS.md
@@ -1,3 +1,8 @@
+# v0.3.2
+
+- improving type inference in `neon_store()` (some more)
+- improved messaging / progress bars for `neon_store()`
+
 # v0.3.1
 
 - improve type inference in `neon_store()`